Simplify UUID generation
In every sensible case a `UUID` only has to change when a new one is
generated, therefore it is useful to return with the same function to
save on the following `get` call.

Furthermore the library function `evalState` can be used to reduce the
work that has to be done at some use sites of the State monad.